requirements.txt
flask
flask-cors
gunicorn
mysql-connector-python
tradingview-ta
yfinance
python-dotenv
numpy
pandas